requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Jira Integration
atlassian-python-api==3.41.16
requests==2.31.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Scheduling
apscheduler==3.10.4
pytz==2024.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Azure OpenAI
openai==1.57.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
itsdangerous==2.1.2

# Rate Limiting
ratelimit==2.2.1

